+++ src/Utils/ImGuiUtils.h
@@ -4,7 +4,7 @@
 
 class ImGuiUtils {
 public:
-	static bool DrawVec2Control(const std::string& label, glm::vec2& values, float resetValue = 0.0f, float columnWidth = 100.0f);
-	static bool DrawVec3Control(const std::string& label, glm::vec3& values, float resetValue = 0.0f, float columnWidth = 100.0f);
-	static bool DrawVec4Control(const std::string& label, glm::vec4& values, float resetValue = 0.0f, float columnWidth = 100.0f);
+	static bool DrawVec2Control(const std::string& label, glm::dvec2& values, float resetValue = 0.0f, float columnWidth = 100.0f);
+	static bool DrawVec3Control(const std::string& label, glm::dvec3& values, float resetValue = 0.0f, float columnWidth = 100.0f);
+	static bool DrawVec4Control(const std::string& label, glm::dvec4& values, float resetValue = 0.0f, float columnWidth = 100.0f);
 };
